+++ CHANGES.md
@@ -25,6 +25,7 @@
 * Lib: make the Wasm version of Json.output work with native ints and JavaScript objects (#1872)
 * Compiler: static evaluation of more primitives (#1912)
 * Compiler: faster compilation by stopping sooner when optimizations become unproductive (#1939)
+* Compiler: improve debug/sourcemap location of closures (#1947)
 * Runtime: use Dataview to convert between floats and bit representation
 
 ## Bug fixes